+++ app/components/SwitchAPI.jsx
@@ -1,15 +1,22 @@
-import React, {Component} from 'react'
-import styles from '../styles/main.css'
-import Select from './material/Select'
+import React, {Component} from 'react';
+import styles from '../styles/main.css';
+import Select from './material/Select';
+import {getErrorText} from '../utils/getConfig';
 
 class SwitchAPI extends Component {
 
     handleChange = (e) => {
         const value = e.target.options[e.target.selectedIndex].value;
         this.props.onBackendChange(value);
-    }
+    };
 
     render() {
+
+        let error_text;
+        if (this.props.error == true) {
+            error_text = getErrorText();
+        }
+
         return (
             <div className={styles.contentcenter}>
                 <h4>Pick an instance</h4>
@@ -18,6 +25,7 @@ class SwitchAPI extends Component {
                         <option value={b.url} key={index}> {b.label} </option>
                     )}
                 </Select>
+                <div className={styles.error_text}>{error_text}</div>
             </div>
         )
     }