+++ app/containers/SwitchAPI.js
@@ -1,6 +1,7 @@
 /*
  * @flow
  */
+
 import { connect } from 'react-redux'
 import { switchBackend, fetchSensors } from '../actions'
 import SwitchAPIComponent from '../components/SwitchAPI'
@@ -10,7 +11,8 @@ import type { Dispatch } from '../utils/flowtype'
 const mapStateToProps = (state) => {
     return {
         selected: state.backends.selected,
-        endpoints: state.backends.endpoints
+        endpoints: state.backends.endpoints,
+        error: state.backends.error
     }
 };